Rename to sky_location module.
