+++ pycbc/distributions/__init__.py
@@ -21,9 +21,9 @@ from pycbc.distributions.angular import *
 from pycbc.distributions.from_file import *
 from pycbc.distributions.gaussian import *
 from pycbc.distributions.power_law import *
+from pycbc.distributions.sky_location import *
 from pycbc.distributions.uniform import *
 from pycbc.distributions.uniform_angle import *
-from pycbc.distributions.uniform_sky import *
 
 # a dict of all available distributions
 distribs = {
