+++ pycbc/distributions/sky_location.py
